add undo redo

+++ primitive/priAnchorWidget.cpp
@@ -8,22 +8,87 @@
 #include "ShapePointGraphicsItem.h"
 #include "AttachTreeNode.h"
 #include "AttachTreeRootNode.h"
+#include "attachTreeNodeMgr.h"
 #include "primitive.h"
 
-namespace pr
+using namespace pr;
+
+PrimitiveAnchorTable::PrimitiveAnchorTable(Primitive *pri, PrimitiveAnchorUi *parent)
+    : QTableWidget(0, 3, parent)
+    , mp_pri(pri)
 {
+    init();
+}
+
+PrimitiveAnchorTable::~PrimitiveAnchorTable()
+{
+    deinit();
+}
+
+void PrimitiveAnchorTable::init()
+{
+    setHorizontalHeaderLabels({tr("X坐标"), tr("Y坐标"), tr("旋转角度")});  // 设置表头
+    setSelectionBehavior(QAbstractItemView::SelectRows);
+    setSelectionMode(QAbstractItemView::SingleSelection);
+    setEditTriggers(QAbstractItemView::DoubleClicked);      // 设置双击编辑触发器
+    horizontalHeader()->setStretchLastSection(true);
+    horizontalHeader()->setSectionResizeMode(QHeaderView::Stretch);
+
+    bool succeed = connect(this, &QTableWidget::itemChanged, this, &PrimitiveAnchorTable::on_item_changed);
+    Q_ASSERT(succeed);
+}
+
+void PrimitiveAnchorTable::deinit()
+{
+    bool succeed = disconnect(this, &QTableWidget::itemChanged, this, &PrimitiveAnchorTable::on_item_changed);
+    Q_ASSERT(succeed);
+}
+
+bool PrimitiveAnchorTable::edit(const QModelIndex &index, EditTrigger trigger, QEvent *event)
+{
+    // 只允许第三列的条目可以编辑
+    if (index.column() == 2)
+    {
+        QTableWidgetItem *item = itemFromIndex(index);
+        if (item)
+        {
+            // 记录原始值
+            m_old_value = item->text();
+        }
+        return QTableWidget::edit(index, trigger, event);
+    }
+    return false;
+}
+
+void PrimitiveAnchorTable::on_item_changed(QTableWidgetItem *item)
+{
+    // 只检查第三列的条目
+    if (item->column() == 2)
+    {
+        QString new_value = item->text();
+        bool is_num = pm::Expression::is_num(new_value);
+        bool is_exp = pm::Expression::isExpression(mp_pri->param_mgr(), new_value);
+
+        // 进行输入值检查
+        if (false == is_num && false == is_exp)
+        {
+            QMessageBox::warning(this, "Warning!", "Your input is invaild!");
+            // 恢复原始值
+            item->setText(m_old_value);
+        }
+        else
+        {
+            emit rotate_angle_changed(item);
+        }
+    }
+}
+
 PrimitiveAnchorUi::PrimitiveAnchorUi(Primitive* pri, QWidget* parent)
     : QWidget(parent)
     , mp_pri(pri)
 {
     QVBoxLayout *layout = new QVBoxLayout();
-    mp_table_widget = new QTableWidget(0, 2);
-    mp_table_widget->setHorizontalHeaderLabels({tr("X坐标"), tr("Y坐标")});
-    mp_table_widget->setSelectionBehavior(QAbstractItemView::SelectRows);
-    mp_table_widget->setSelectionMode(QAbstractItemView::SingleSelection);
-    mp_table_widget->setEditTriggers(QAbstractItemView::NoEditTriggers);
-    mp_table_widget->horizontalHeader()->setStretchLastSection(true);
-    mp_table_widget->horizontalHeader()->setSectionResizeMode(QHeaderView::Stretch);
+    mp_table_widget = new PrimitiveAnchorTable(mp_pri, this); // 进行内存托管
 
     QHBoxLayout *button_bottom = new QHBoxLayout();
     mp_del_button = new QPushButton(QIcon(":/img/lay_del.png"), tr("删除"));
@@ -40,9 +105,9 @@ PrimitiveAnchorUi::PrimitiveAnchorUi(Primitive* pri, QWidget* parent)
     button_bottom->addWidget(mp_del_button);
 
     // 选中item connect
-    connect(mp_table_widget, &QTableWidget::itemClicked, this, &PrimitiveAnchorUi::onItemSelectionChanged);
-
-    connect(mp_del_button, &QPushButton::clicked, this, &PrimitiveAnchorUi::onDeleteItem);
+    connect(mp_table_widget, &QTableWidget::itemClicked, this, &PrimitiveAnchorUi::on_item_selection_changed);
+    connect(mp_table_widget, &PrimitiveAnchorTable::rotate_angle_changed, this, &PrimitiveAnchorUi::on_rotate_angle_changed);
+    connect(mp_del_button, &QPushButton::clicked, this, &PrimitiveAnchorUi::on_delete_item);
 
     layout->addWidget(mp_table_widget);
     layout->addLayout(button_bottom);
@@ -54,7 +119,37 @@ PrimitiveAnchorUi::~PrimitiveAnchorUi()
 {
 }
 
-void PrimitiveAnchorUi::addAnchorToTable(const QVariant &data)
+void PrimitiveAnchorUi::save_anchors_to_primitive()
+{
+    QVector<at::AttachTreeUtils::AttachPointPosInf> anchors;
+    for(auto itor = m_anchor_list.begin(); itor != m_anchor_list.end(); itor++)
+    {
+        auto point_item = itor->first;
+        auto tree_node = dynamic_cast<at::AttachTreeNode*>(at::AttachTreeUtils::attach_tree_node_point_item_in(point_item, mp_pri->at_root()));
+        int idx = at::AttachTreeUtils::point_item_index_in_tree_node(tree_node, point_item);
+        pm::PointE point_ex = at::AttachTreeUtils::attach_exp_point_coord(tree_node, idx, true);
+        at::AttachTreeUtils::AttachPointPosInf point_info = {tree_node->id(), idx, point_ex, point_item->rotate_angle()};
+        anchors.append(point_info);
+    }
+    mp_pri->set_json_anchors(anchors);
+}
+
+void PrimitiveAnchorUi::load_anchors_from_primitive()
+{
+    QVector<at::AttachTreeUtils::AttachPointPosInf> anchors = mp_pri->get_json_anchors();
+    for(auto itor = anchors.begin(); itor != anchors.end(); itor++)
+    {
+        auto tree_node = mp_pri->at_root()->tree_node_mgr()->query(itor->id);
+        auto idx = itor->attach_point_idx;
+        auto is_shape = true;
+        tree_node->point_items()[idx]->set_rotate_angle(itor->rotate_angle);
+        TreeNodeData node_data{tree_node, idx, is_shape};
+        QVariant variant = QVariant::fromValue(node_data);
+        add_anchor_to_table(variant);
+    }
+}
+
+void PrimitiveAnchorUi::add_anchor_to_table(const QVariant &data)
 {
     // QVariant转为treeNodeData
     TreeNodeData node_data = data.value<TreeNodeData>();
@@ -67,34 +162,48 @@ void PrimitiveAnchorUi::addAnchorToTable(const QVariant &data)
     //获取表达式
     pm::PointE exp_point("0", "0") ;
     auto tree_node = dynamic_cast<at::AttachTreeNode*>(node_data.node_pointer);
+    sp::ShapePointGraphicsItem* point_item = nullptr;
     if(tree_node)
     {
-        exp_point = at::AttachTreeUtils::attach_exp_point(tree_node, node_data.point_idx_in_shape, true);
+        exp_point = at::AttachTreeUtils::attach_exp_point_coord(tree_node, node_data.point_idx_in_shape, true);
+        point_item = tree_node->point_items()[node_data.point_idx_in_shape];
+    }
+    else
+    {
+        point_item = mp_pri->at_root()->origin_point();
     }
 
+    if(is_point_item_existed(point_item))
+    {
+        QMessageBox::warning(nullptr, "Warning!", "Current attach point has been anchor point!");
+        return;
+    }
     //将attach point添加到table widget中
     int rowCount = mp_table_widget->rowCount();
     mp_table_widget->insertRow(rowCount);
     QTableWidgetItem *x_item = new QTableWidgetItem(QString("【%1】: %2").arg(exp_point.x().to_double(mp_pri->param_mgr())).arg(exp_point.x().to_str()));
     QTableWidgetItem *y_item = new QTableWidgetItem(QString("【%1】: %2").arg(exp_point.y().to_double(mp_pri->param_mgr())).arg(exp_point.y().to_str()));
+    QTableWidgetItem *rotate_angle_item = new QTableWidgetItem(point_item->rotate_angle());
     //mp_table_widget->setItem(rowCount, 0, new QTableWidgetItem(QString::number(m_anchor_num++)));
     mp_table_widget->setItem(rowCount, 0, x_item);
     mp_table_widget->setItem(rowCount, 1, y_item);
+    mp_table_widget->setItem(rowCount, 2, rotate_angle_item);
 
     //将treenode信息和point_idx添加到x_item的data中
-    x_item->setData(Qt::UserRole, QVariant::fromValue(data));
+    rotate_angle_item->setData(Qt::UserRole, QVariant::fromValue(data));
 
     //将锚点的treenodeid和rowCount对应关系保存起来
-    m_anchor_map.insert(QPair<at::AttachTreeNodeId, int>(tree_node->id(), node_data.point_idx_in_shape), x_item);
+    QPair<sp::ShapePointGraphicsItem*, QTableWidgetItem*> pair{point_item, rotate_angle_item};
+    m_anchor_list.append(pair);
 }
 
-void PrimitiveAnchorUi::onDeleteItem()
+void PrimitiveAnchorUi::on_delete_item()
 {
     // 获取选中的行
     int row = mp_table_widget->currentRow();
 
     // 获取选中行中的data数据
-    QTableWidgetItem *item = mp_table_widget->item(row, 0);
+    QTableWidgetItem *item = mp_table_widget->item(row, 2);
     if (item)
     {
         // 获取data数据
@@ -106,14 +215,21 @@ void PrimitiveAnchorUi::onDeleteItem()
         }
 
         // 将该附着点设置为非锚点
-
         auto tree_node = dynamic_cast<at::AttachTreeNode*>(node_data.node_pointer);
         if(tree_node)
         {
-            tree_node->set_anchor_point(node_data.point_idx_in_shape, false);
-
+            auto point_item = tree_node->point_items()[node_data.point_idx_in_shape];
+            point_item->set_anchor_point(false);
+            point_item->set_rotate_angle("0");
             //删除锚点的treenodeid和rowCount对应关系
-            m_anchor_map.remove(QPair<at::AttachTreeNodeId, int>(tree_node->id(), node_data.point_idx_in_shape));
+            for(auto itor = m_anchor_list.begin(); itor < m_anchor_list.end(); itor++)
+            {
+                if(itor->first == point_item)
+                {
+                    m_anchor_list.erase(itor);
+                    break;
+                }
+            }
         }
         else
         {
@@ -125,13 +241,13 @@ void PrimitiveAnchorUi::onDeleteItem()
     mp_table_widget->removeRow(row);
 }
 
-void PrimitiveAnchorUi::onItemSelectionChanged()
+void PrimitiveAnchorUi::on_item_selection_changed()
 {
     // 获取选中的行
     int row = mp_table_widget->currentRow();
 
     // 获取选中行中的data数据
-    QTableWidgetItem *item = mp_table_widget->item(row, 0);
+    QTableWidgetItem *item = mp_table_widget->item(row, 2);
     if (item)
     {
         // 获取data数据
@@ -154,19 +270,39 @@ void PrimitiveAnchorUi::onItemSelectionChanged()
     }
 }
 
+void PrimitiveAnchorUi::on_rotate_angle_changed(QTableWidgetItem *item)
+{
+    for(auto itor = m_anchor_list.begin(); itor != m_anchor_list.end(); itor++)
+    {
+        if(itor->second == item)
+        {
+            itor->first->set_rotate_angle(item->text());
+            break;
+        }
+    }
+}
+
 void PrimitiveAnchorUi::insert_an_item(at::AttachTreeNode* tree_node, int attach_point_idx)
 {
     //获取表达式
     pm::PointE exp_point("0", "0") ;
-    exp_point = at::AttachTreeUtils::attach_exp_point(tree_node, attach_point_idx, true);
+    exp_point = at::AttachTreeUtils::attach_exp_point_coord(tree_node, attach_point_idx, true);
 
+    auto point_item = tree_node->point_items()[attach_point_idx];
+    if(is_point_item_existed(point_item))
+    {
+        QMessageBox::warning(nullptr, "Warning!", "Current attach point has been anchor point!");
+        return;
+    }
     //将attach point添加到table widget中
     int rowCount = mp_table_widget->rowCount();
     mp_table_widget->insertRow(rowCount);
     QTableWidgetItem *x_item = new QTableWidgetItem(QString("【%1】: %2").arg(exp_point.x().to_double(mp_pri->param_mgr())).arg(exp_point.x().to_str()));
     QTableWidgetItem *y_item = new QTableWidgetItem(QString("【%1】: %2").arg(exp_point.y().to_double(mp_pri->param_mgr())).arg(exp_point.y().to_str()));
+    QTableWidgetItem *rotate_angle_item = new QTableWidgetItem(point_item->rotate_angle());
     mp_table_widget->setItem(rowCount, 0, x_item);
     mp_table_widget->setItem(rowCount, 1, y_item);
+    mp_table_widget->setItem(rowCount, 2, rotate_angle_item);
 
     TreeNodeData data;
     data.node_pointer       = tree_node;
@@ -174,10 +310,22 @@ void PrimitiveAnchorUi::insert_an_item(at::AttachTreeNode* tree_node, int attach
     data.point_idx_in_shape = attach_point_idx;
 
     //将treenode信息和point_idx添加到x_item的data中
-    x_item->setData(Qt::UserRole, QVariant::fromValue(data));
-
+    rotate_angle_item->setData(Qt::UserRole, QVariant::fromValue(data));
+    QPair<sp::ShapePointGraphicsItem*, QTableWidgetItem*> pair{point_item, rotate_angle_item};
     //将锚点的treenodeid和rowCount对应关系保存起来
-    m_anchor_map.insert(QPair<at::AttachTreeNodeId, int>(tree_node->id(), data.point_idx_in_shape), x_item);
+    m_anchor_list.append(pair);
+}
+
+bool PrimitiveAnchorUi::is_point_item_existed(sp::ShapePointGraphicsItem *point_item)
+{
+    for(auto pair : m_anchor_list)
+    {
+        if(pair.first == point_item)
+        {
+            return true;
+        }
+    }
+    return false;
 }
 
 void PrimitiveAnchorUi::clear_table()
@@ -187,61 +335,39 @@ void PrimitiveAnchorUi::clear_table()
     mp_table_widget->setRowCount(0);
 
     // 清空锚点的treenodeid和rowCount对应关系
-    m_anchor_map.clear();
+    for(auto pair : m_anchor_list)
+    {
+        pair.first->set_anchor_point(false);
+        pair.first->set_rotate_angle("0");
+    }
+    m_anchor_list.clear();
 }
 
 void PrimitiveAnchorUi::update_table()
 {
-    // 清空列表
-    clear_table();
-
-    // 通过读取信息重建列表
-    rebuild_anchor_table(mp_pri->at_root());
-}
-
-void PrimitiveAnchorUi::rebuild_anchor_table(at::AttachTreeBaseNode* root_node)
-{
-    // 使用TreeNode的递归遍历，这里实现对节点的处理
-    root_node->traversal_subtree([this](at::AttachTreeBaseNode* node)
-    {
-        // 根节点不做处理
-        if(dynamic_cast<at::AttachTreeRootNode * >(node))
-        {
-            return;
-        }
-
-        // 附着点如果是锚点，插入到列表中
-        at::AttachTreeNode* attach_tree_node = dynamic_cast<at::AttachTreeNode*>(node);
-        if(nullptr == attach_tree_node)
-        {
-            return;
-        }
-        for(int attach_idx = 0; attach_idx < attach_tree_node->point_items().size(); attach_idx++)
-        {
-            if(attach_tree_node->is_anchor_point(attach_idx))
-            {
-                insert_an_item(attach_tree_node, attach_idx);
-            }
-        }
-    });
+    load_anchors_from_primitive();
 }
 
 // 外部删除节点后触发的槽函数
-void PrimitiveAnchorUi::onDeleteNode(at::AttachTreeNode* tree_node)
+void PrimitiveAnchorUi::on_delete_node(at::AttachTreeNode* tree_node)
 {
-    tree_node->traversal_subtree([=](at::AttachTreeBaseNode* node){
+    tree_node->traversal_subtree([ = ](at::AttachTreeBaseNode* node)
+    {
         at::AttachTreeNode* sub_node = dynamic_cast<at::AttachTreeNode* >(node);
-        for(auto idx = 0; idx < sub_node->point_items().size(); idx++)
+        auto point_items = sub_node->point_items();
+        for(auto idx = 0; idx < point_items.size(); idx++)
         {
+            auto& point_item = point_items[idx];
             // 查找该点是否记录在m_anchor_map中，如果是，则删除该点
-            auto it = m_anchor_map.find(QPair<at::AttachTreeNodeId, int>(sub_node->id(), idx));
-            if (it != m_anchor_map.end())
+            for(auto itor = m_anchor_list.begin(); itor != m_anchor_list.end(); itor++)
             {
-                // 如果是，则删除该点
-                mp_table_widget->removeRow(mp_table_widget->row(it.value()));
-                m_anchor_map.erase(it);
+                if(itor->first == point_item)
+                {
+                    mp_table_widget->removeRow(mp_table_widget->row(itor->second));
+                    m_anchor_list.erase(itor);
+                    break;
+                }
             }
         }
     });
-}
 }
